Separated reducers into their own files
