+++ client/src/App.js
@@ -3,8 +3,8 @@ import './App.scss';
 import UserProfile_LI from './components/UserProfile_LI';
 import { Route, Switch } from 'react-router-dom';
 import { useAuth0 } from './auth-wrapper.js';
-import './App.css';
-import NavBar from './components/NavBar.js';
+import './App.scss';
+import Navbar from './components/Navbar.js';
 import FakeProfile from './components/FakeProfile.js';
 import PrivateRoute from './components/PrivateRoute.js';
 
@@ -17,7 +17,7 @@ function App() {
 
   return (
     <div className="App">
-      <NavBar />
+      <Navbar />
       <Switch>
         <Route exact path="/" />
         <Route exact path="/profile" component={UserProfile_LI} />
