+++ client/src/store/actions/index.js
@@ -91,3 +91,7 @@ export const getSingleUser = id => dispatch => {
 export const UPDATE_USER_START = 'UPDATE_USER_START';
 export const UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS';
 export const UPDATE_USER_FAILURE = 'UPDATE_USER_FAILURE';
+
+export const GET_SINGLE_PROJECT_START = 'GET_SINGLE_PROJECT_START';
+export const GET_SINGLE_PROJECT_SUCCESS = 'GET_SINGLE_PROJECT_SUCCESS';
+export const GET_SINGLE_PROJECT_FAILURE = 'GET_SINGLE_PROJECT_FAILURE';
